backend/requirements.txt
# FastAPI and server
fastapi[standard]==0.115.12
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# AI/ML
httpx>=0.25.0
sentence-transformers>=3.0.0
numpy>=1.24.0
openai-whisper>=20231117
soundfile>=0.12.1
scipy>=1.11.0
librosa>=0.10.0
resampy>=0.4.2
scikit-learn>=1.3.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.1.0

# PyTorch (required for embeddings service and TTS)
torch>=2.0.0
torchaudio>=2.0.0

# TTS - piper-tts integration
piper-tts

# WebSocket and async
websockets>=12.0
aiofiles>=23.2.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0

# Authentication & Security
bcrypt>=4.0.0

# System integration (for hotkeys)
pynput>=1.7.6
pyaudio>=0.2.13

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0